Tried to explicit with dtypes to avoid issues with windows vs unix

+++ gaitmap/data_transform/_feature_transform.py
@@ -116,7 +116,7 @@ class Resample(BaseTransformer):
             self.roi_list = roi_list
             out = roi_list.copy()
             out.loc[:, ["start", "end"]] = (
-                (roi_list[["start", "end"]] * self.target_sampling_rate_hz / self.sampling_rate_hz).round().astype(int)
+                (roi_list[["start", "end"]] * self.target_sampling_rate_hz / self.sampling_rate_hz).round()
             )
             self.transformed_roi_list_ = out
         return self
