+++ gaitmap_mad/gaitmap_mad/gait_detection/_ullrich_gait_sequence_detection.py
@@ -378,14 +378,14 @@ class UllrichGaitSequenceDetection(BaseGaitDetection):
         #  implementation would have found the second peak and correctly concluded that the harmonic was found.
         #  With this implementation method, we find the first and then conclude that the peak prominence is too low and
         #  discard it.
-        closest_peaks = find_extrema_in_radius(f_s_1d_flat, harmonics_flat, harmonic_window_half, "max").astype(int)
+        closest_peaks = find_extrema_in_radius(f_s_1d_flat, harmonics_flat, harmonic_window_half, "max").astype("int32")
 
         peak_prominence = peak_prominences(f_s_1d_flat, closest_peaks)[0].reshape(harmonics_candidates.shape)
         peak_heights = f_s_1d_flat[closest_peaks].reshape(harmonics_candidates.shape)
 
         # Apply thresholds
         # peaks should be higher than mean of f_s_1d in the area <= lp_freq_hz. Have to add a + 1 to include the limit
-        min_peak_height = np.mean(f_s_1d[:, : np.floor(lp_freq_hz / freq_axis_delta).astype(int) + 1], axis=1)
+        min_peak_height = np.mean(f_s_1d[:, : np.floor(lp_freq_hz / freq_axis_delta).astype("int32") + 1], axis=1)
         # duplicate to match the shape of peak_heights
         min_peak_height = np.tile(min_peak_height, (peak_heights.shape[1], 1)).T
